pyproject.toml
[project]
name = "monstercatcher"
version = "0.1.0"
description = "Archived monster-catching engine experiment"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pygame-ce>=2.5.3",
]

[project.scripts]
monstercatcher = "monstercatcher:main"

[build-system]
requires = ["setuptools>=82.0.1"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["monstercatcher"]

[tool.setuptools.packages.find]
where = ["src"]
